+++ app/models/form.rb
@@ -1,13 +1,24 @@
+# Represents the PDF or CRF on which a variable was captured
 class Form < ActiveRecord::Base
-
   # Model Validation
-  validates_presence_of :name, :dataset_id
-  validates_format_of :name, with: /\A[a-z]\w*\Z/i
-  validates_uniqueness_of :name, scope: :dataset_id
+  validates :name, :dataset_id, presence: true
+  validates :name, format: { with: /\A[a-z]\w*\Z/i }
+  validates :name, uniqueness: { scope: :dataset_id }
 
   # Model Relationships
   belongs_to :dataset
   has_many :variable_forms
   has_many :variables
 
+  def pdf?
+    extension == 'pdf'
+  end
+
+  # def image?
+  #   %w(png jpg jpeg gif).include?(extension)
+  # end
+
+  def extension
+    code_book.split('.').last.to_s.downcase
+  end
 end