+++ CHANGELOG.md
@@ -12,6 +12,7 @@
   - Variables are now sorted by relevance when searching for key terms
   - Individual variable pages have been redesigned and include additional information about known issues and variable history
   - Folder navigation has been simplified on the variable index
+  - Embedded PDFs for variables are now displayed directly on the variable page
 - **Gem Changes**
   - Updated to rails 4.2.5
   - Updated to pg 0.18.4